handlers/menu.py
# handlers/menu.py
import logging
from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import ContextTypes # Removed CommandHandler, CallbackQueryHandler, ConversationHandler if not a conv

import config # For states and callback prefixes
# Import entry points of other handlers if menu directly triggers them
# from . import create, update, delete, read, insights # These are not directly called by main_menu_command_simple

logger = logging.getLogger(__name__)

async def main_menu_command_simple(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    """Displays the main menu with inline buttons."""
    # This function was provided in a previous response. Ensure it's correct.
    keyboard = [
        [InlineKeyboardButton("➕ New Entry", callback_data=f"{config.CB_MAIN_MENU_PREFIX}new")],
        [InlineKeyboardButton("✏️ Update Entry", callback_data=f"{config.CB_MAIN_MENU_PREFIX}update")],
        [InlineKeyboardButton("🗑️ Delete Entry", callback_data=f"{config.CB_MAIN_MENU_PREFIX}delete")],
        [InlineKeyboardButton("👁️ View Entries", callback_data=f"{config.CB_MAIN_MENU_PREFIX}view")],
        [InlineKeyboardButton("📊 Price Insights", callback_data=f"{config.CB_MAIN_MENU_PREFIX}insights")],
    ]
    reply_markup = InlineKeyboardMarkup(keyboard)
    message_text = "Welcome! Select an option:" # Or your preferred menu text
    
    origin_message = update.message or update.callback_query.message # Get the message object
    
    if update.callback_query: # If called from another menu or button press
        await update.callback_query.answer()
        # Edit the message that contained the button
        await origin_message.edit_text(text=message_text, reply_markup=reply_markup)
    elif update.message: # If called by a command like /menu
        await origin_message.reply_text(text=message_text, reply_markup=reply_markup)
    else:
        logger.error("main_menu_command_simple called without message or callback_query in update.")

# Any other menu-related functions you might have...